use libswscale replacing media foundation

+++ Tomato.ASIK.COM/MFIOProvider.h
@@ -69,7 +69,7 @@ private:
 	Tomato::block_buffer<byte> buffer;
 	std::unique_ptr<short[]> samples;
 	std::unique_ptr<Tomato::ASIK::Core::spectrogram> spectr;
-	std::vector<uint32_t> specData;
+	std::vector<byte> specData;
 	size_t img_Width, img_Height;
 	size_t samples_count;
 public:
